Revert "Add Location header to finalize response (#85)" (#509)
This reverts commit 9c3d91bf5b6b798a677d50232e6149ab99620a52, which was
merged as a result of https://github.com/letsencrypt/pebble/pull/85.
That PR based on the idea that the Location header is required to be
included with the Finalize response in RFC 8555, but this is not the
case: it appears in the example response, but its inclusion there is not
mandated by the text.

Although Boulder does provide the Location header in Finalize responses,
there exist other ACME implementations which do not. Therefore it is
useful for Pebble to omit it, so that clients developed against Pebble
do not come to rely on its inclusion.

Fixes https://github.com/letsencrypt/pebble/issues/508