Allow generating ECDSA roots and intermediates (#518)
Add a new "keyAlgorithm" entry to Pebble's config, which can take the
value "rsa" or "ecdsa". Leaving it unset or setting it to "rsa"
preserves the same behavior as today, in which Pebble uses RSA 2048 to
generate all of its root and intermediate keys and certificates. Setting
it to "ecdsa" causes Pebble to use ECDSA P-256 for all of its root and
intermediate keys and certificates.

This paves the way for having pebble support other
not-approved-by-the-BRs key algorithms, such as ML-DSA.